Initial ingredients refactor
