+++ .http/README.md
@@ -1,0 +1,7 @@
+# API Testing
+
+## Setup
+Set secrets in Environments → local:
+- `client_id` - Your client ID
+- `client_secret` - Your client secret
+- `audience` - API audience
