+++ .http/prep-api/environments/local.bru
@@ -1,10 +1,10 @@
 vars {
-  audience: test
   auth_url: https://dev-43j7zsnkz5026sb5.us.auth0.com
   api_base_url: http://localhost:5053
 }
 vars:secret [
   client_id,
   client_secret,
+  audience,
   access_token
 ]
