+++ src/PrepApi/Recipes/Requests/CreateVariantFromPrepRequest.cs
@@ -2,6 +2,6 @@ namespace PrepApi.Recipes.Requests;
 
 public class CreateVariantFromPrepRequest
 {
-    public required string Name { get; set; }
-    public bool SetAsFavorite { get; set; }
+    public required string Name { get; init; }
+    public bool SetAsFavorite { get; init; }
 }
