+++ src/PrepApi/Users/User.cs
@@ -5,7 +5,7 @@ namespace PrepApi.Users;
 public class User : Entity
 {
     public required string ExternalId { get; init; }
-    public string? Email { get; set; }
+    public string? Email { get; init; }
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
     public PreferredUnits PreferredUnits { get; set; } = PreferredUnits.Metric;
